# src/Prdb.Fab.Host/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "prdb-fab",
    "description": "The API the browser side of prdb-fab talks to. Generated from the code at build time and committed: change an endpoint, not this file.",
    "version": "v1"
  },
  "paths": {
...
    "/api/catalogue/whats-new": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhatsNewPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalogue/whats-new/observed": {
      "post": {
        "tags": [
          "Catalogue"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatsNewObservation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/catalogue/wanted": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WantedList"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalogue/videos": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalogue/wanted/{prdbId}": {
      "post": {
        "tags": [
          "Catalogue"
        ],
        "parameters": [
          {
            "name": "prdbId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountPreferenceVerdict"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Catalogue"
        ],
        "parameters": [
          {
            "name": "prdbId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountPreferenceVerdict"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalogue/actors/{prdbId}/favourite": {
      "post": {
        "tags": [
          "Catalogue"
        ],
        "parameters": [
          {
            "name": "prdbId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountPreferenceVerdict"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Catalogue"
        ],
        "parameters": [
          {
            "name": "prdbId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountPreferenceVerdict"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalogue/sites/{prdbId}/favourite": {
      "post": {
        "tags": [
          "Catalogue"
        ],
        "parameters": [
          {
            "name": "prdbId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountPreferenceVerdict"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Catalogue"
        ],
        "parameters": [
          {
            "name": "prdbId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountPreferenceVerdict"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalogue/videos/{prdbId}/download-best": {
      "post": {
        "tags": [
          "Catalogue"
        ],
        "parameters": [
          {
            "name": "prdbId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DownloadVerdict"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/catalogue/sites": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "scope",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SitePage"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalogue/sites/{prdbId}": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "parameters": [
          {
            "name": "prdbId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteVideos"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/catalogue/actors": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "scope",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorPage"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalogue/actors/{prdbId}": {
      "get": {
        "tags": [
          "Catalogue"
        ],
        "parameters": [
          {
            "name": "prdbId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32",
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorVideos"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountPreferenceOutcome": {
        "enum": [
          "Updated",
          "Deferred",
          "Rejected",
          "NotFound",
          "Failed"
        ]
      },
      "AccountPreferenceVerdict": {
        "required": [
          "outcome",
          "desired",
          "detail"
        ],
        "type": "object",
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/AccountPreferenceOutcome"
          },
          "desired": {
            "type": "boolean"
          },
          "detail": {
            "type": "string"
          },
          "retryAfterSeconds": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "updated": {
            "type": "boolean"
          }
        }
      },
      "ActorCard": {
        "required": [
          "prdbId",
          "name",
          "videoCount",
          "favourite",
          "hasArtwork"
        ],
        "type": "object",
        "properties": {
          "prdbId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "videoCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "favourite": {
            "type": "boolean"
          },
          "hasArtwork": {
            "type": "boolean"
          }
        }
      },
      "ActorPage": {
        "required": [
          "actors",
          "page",
          "pageSize",
          "total",
          "scope"
        ],
        "type": "object",
        "properties": {
          "actors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActorCard"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "scope": {
            "$ref": "#/components/schemas/CatalogueScope"
          }
        }
      },
      "ActorVideos": {
        "required": [
          "actor",
          "videos"
        ],
        "type": "object",
        "properties": {
          "actor": {
            "$ref": "#/components/schemas/BrowseContext"
          },
          "videos": {
            "$ref": "#/components/schemas/VideoPage"
          }
        }
      },
...
      "BrowseContext": {
        "required": [
          "prdbId",
          "title",
          "favourite"
        ],
        "type": "object",
        "properties": {
          "prdbId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "favourite": {
            "type": "boolean"
          }
        }
      },
      "CatalogueScope": {
        "enum": [
          "Favourites",
          "All"
        ]
      },
...
      "DownloadCause": {
        "enum": [
          "Rejected",
          "Failed",
          "Unusable",
          "Vanished",
          "Abandoned",
          "Empty",
          null
        ]
      },
...
      "DownloadOutcome": {
        "enum": [
          "Pending",
          "Submitted",
          "Rejected",
          "SubmissionUnknown",
          "ConnectionProblem",
          "IndexerProblem",
          "ReleaseNotEligible",
          "NoReleasesLeft",
          "RetryBudgetSpent"
        ]
      },
...
      "DownloadState": {
        "enum": [
          "Outstanding",
          "Completed",
          "Collected",
          "Failed",
          "Abandoned"
        ]
      },
      "DownloadVerdict": {
        "required": [
          "outcome",
          "downloadId",
          "state",
          "cause",
          "nzoId",
          "detail"
        ],
        "type": "object",
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/DownloadOutcome"
          },
          "downloadId": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DownloadState"
              }
            ]
          },
          "cause": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DownloadCause"
              }
            ]
          },
          "nzoId": {
            "type": [
              "null",
              "string"
            ]
          },
          "detail": {
            "type": "string"
          }
        }
      },
...
      "SiteCard": {
        "required": [
          "prdbId",
          "title",
          "network",
          "videoCount",
          "favourite",
          "representativeVideoId"
        ],
        "type": "object",
        "properties": {
          "prdbId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "network": {
            "type": [
              "null",
              "string"
            ]
          },
          "videoCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "favourite": {
            "type": "boolean"
          },
          "representativeVideoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          }
        }
      },
...
      "SitePage": {
        "required": [
          "sites",
          "page",
          "pageSize",
          "total",
          "scope"
        ],
        "type": "object",
        "properties": {
          "sites": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SiteCard"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "scope": {
            "$ref": "#/components/schemas/CatalogueScope"
          }
        }
      },
      "SiteVideos": {
        "required": [
          "site",
          "videos"
        ],
        "type": "object",
        "properties": {
          "site": {
            "$ref": "#/components/schemas/BrowseContext"
          },
          "videos": {
            "$ref": "#/components/schemas/VideoPage"
          }
        }
      },
...
      "VideoAvailability": {
        "enum": [
          "Ready",
          "ReleasesNeedInspection",
          "NoIdentifiedRelease"
        ],
        "default": "NoIdentifiedRelease"
      },
      "VideoCard": {
        "required": [
          "id",
          "prdbId",
          "title",
          "site",
          "sitePrdbId",
          "releaseDate"
        ],
        "type": "object",
        "properties": {
          "id": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "prdbId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "site": {
            "type": [
              "null",
              "string"
            ]
          },
          "sitePrdbId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "releaseDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "downloadReady": {
            "type": "boolean",
            "default": false
          },
          "wanted": {
            "type": "boolean",
            "default": false
          },
          "wantedSyncPending": {
            "type": "boolean",
            "default": false
          },
          "wantedSyncFailure": {
            "type": [
              "null",
              "string"
            ]
          },
          "outstanding": {
            "type": "boolean",
            "default": false
          },
          "heldQualities": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "availability": {
            "$ref": "#/components/schemas/VideoAvailability"
          }
        }
      },
      "VideoPage": {
        "required": [
          "videos",
          "page",
          "pageSize",
          "total"
        ],
        "type": "object",
        "properties": {
          "videos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoCard"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "WantedList": {
        "required": [
          "videos",
          "page",
          "pageSize",
          "total",
          "feedHasRun",
          "backfillRunning"
        ],
        "type": "object",
        "properties": {
          "videos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoCard"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "feedHasRun": {
            "type": "boolean"
          },
          "backfillRunning": {
            "type": "boolean"
          }
        }
      },
      "WhatsNewObservation": {
        "required": [
          "videoId",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "videoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "WhatsNewPage": {
        "required": [
          "videos",
          "page",
          "pageSize",
          "total",
          "newCount",
          "checkpointVideoId",
          "checkpointCreatedAt"
        ],
        "type": "object",
        "properties": {
          "videos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VideoCard"
            }
          },
          "page": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "pageSize": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "newCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "checkpointVideoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "checkpointCreatedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      }
    }
  },
...
}